[RELEASE=0.1.171] runner now pulls images onto nodes.
runner.run_cyber_dojo_sh() now returns colour==pulling and
asynchronously pulls the image, when the image_name in the
manifest is not present on the node.

web shows new dialog when [test] returns 'pulling', saying
the image is not ready but will be very shortly.

nginx has a new pulling 'traffic-light' hour-glass icon
for this [test] result.